Refactor: move `LagoApiError` to `exceptions` module
