+++ lago_python_client/clients/base_client.py
@@ -1,5 +1,3 @@
-from http import HTTPStatus
-import sys
 from typing import Any, Optional
 from urllib.parse import urljoin, urlencode
 
@@ -7,13 +5,9 @@ from pydantic import BaseModel
 import requests
 from requests import Response
 
-from lago_python_client.version import LAGO_VERSION
+from ..exceptions import LagoApiError
 from ..services.json import from_json, to_json
-
-if sys.version_info < (3, 9):
-    from typing import MutableMapping
-else:
-    from collections.abc import MutableMapping
+from ..version import LAGO_VERSION
 
 
 class BaseClient:
@@ -130,25 +124,3 @@ class BaseClient:
         }
 
         return response
-
-
-class LagoApiError(Exception):
-    def __init__(
-        self,
-        status_code: int,
-        url: Optional[str],
-        response: Any,
-        detail: Optional[str] = None,
-        headers: Optional[MutableMapping[str, str]] = None,
-    ) -> None:
-        if detail is None:
-            detail = HTTPStatus(status_code).phrase
-        self.status_code = status_code
-        self.url = url
-        self.response = response
-        self.detail = detail
-        self.headers = headers
-
-    def __repr__(self) -> str:
-        class_name = self.__class__.__name__
-        return f"{class_name}(status_code={self.status_code!r}, detail={self.detail!r})"
