+++ tests/test_credit_note_client.py
@@ -3,9 +3,9 @@ import requests_mock
 import os
 
 from lago_python_client.client import Client
+from lago_python_client.exceptions import LagoApiError
 from lago_python_client.models.invoice import FeeResponse
 from lago_python_client.models.credit_note import Item, Items, CreditNote, CreditNoteUpdate
-from lago_python_client.clients.base_client import LagoApiError
 
 def credit_note_object():
     item1 = Item(
