+++ tests/test_plan_client.py
@@ -3,8 +3,8 @@ import requests_mock
 import os
 
 from lago_python_client.client import Client
+from lago_python_client.exceptions import LagoApiError
 from lago_python_client.models import Plan, Charge, Charges
-from lago_python_client.clients.base_client import LagoApiError
 
 
 def plan_object():
