+++ tests/test_wallet_client.py
@@ -3,8 +3,8 @@ import requests_mock
 import os
 
 from lago_python_client.client import Client
+from lago_python_client.exceptions import LagoApiError
 from lago_python_client.models import Wallet
-from lago_python_client.clients.base_client import LagoApiError
 
 
 def wallet_object():
